# Speed up large selections in the editor