There are many places where selections have `O(n^2)` behavior, likely because they're generally small enough for it not to matter. When trying to work with large scenes, however, I naturally found myself selecting 1000 or more elements at a time, and that's around the scale where that sort of behavior starts to blow up. 

I realize most people aren't going to be working with selections this large, but I don't see a good reason not to support it. Some of these changes slightly slow down small selections, some speed them up. I didn't bother testing whether small selections were sped up or slowed down on the net because they're still fast and infrequent enough for it not to matter.

I've split this PR in two commits to make it easier to review; I'll squash them once people are happy. About half the changes are just using references instead of copies, and it clutters up the actually interesting changes, so I split them out into their own commit.

Most of these changes are direct fixes to issues that showed up on a flamegraph while I was testing, so they do have actual impact. The exception is the swapping of copies for const references. After I ran into two instances that mattered, I just searched the entire codebase and fixed them all, so some of them probably don't matter in practice.

Most of the changes are just complexity class improvements that should be self explanatory. The `editor_data` change is a fix for what looks like a rather impactful typo. The two signal emission updates prevent doing work until the selection is done changing.

This speeds up many types of selections and deselections.

Selecting 5000 sprites via the scene tree dock: 6s -> ~0.25s
Deselecting 5000 sprites by selecting 1 new sprite in the scene tree dock: 4s -> imperceptible
Selecting 25000 sprites via the scene tree dock: 172s -> ~0.5s

There's similarly massive speedups to selection via drag select and to more types than just sprites, but you basically know what to expect from the above numbers.

---------------------

All performance tests were run on builds made with the following command: `scons optimize=speed_trace debug_symbols=yes platform=linuxbsd dev_build=no production=yes`.

The test project [editor-optimization-test-scenes.zip](https://github.com/user-attachments/files/21709465/editor-optimization-test-scenes.zip) consists of a series of scenes that each have 25,000 of a common type of node. Unless you also apply TODO edit in, it might be difficult to test anything but the `Sprite2D.tscn` scene.