@KoBeWi 2025-12-02
#111837 should fix it.